+++ cc/82_remove_duplicates_from_sorted_list_ii.cc
@@ -79,7 +79,7 @@ void ExpectEqList(ListNode* const head, const std::vector<int>& elements) {
         m_head = m_head->next;
         i++;
     }
-    EXPECT_EQ(i, elements.size());
+    EXPECT_EQ(i, static_cast<int>(elements.size()));
 }
 
 TEST(Test, t1) {
